Move DEXPREOPT down to device tree.
This is a waste of cpu-time, life-time and bamdwith; and as soon
as you wipe dalvik and cache after flash, the advantage is gone again.

For my building host the preopting take 17 minutes longer - this 17
minutes are even used, if the complete rom is build, and you just
rebuild for a change in sepolicy...

Short version: for me preopting has much more disadvantages then
advanteges.

Change-Id: I3042911bf425bdb23ac87b30d3d52085c37ab850